+++ packages/jsx-scanner/src/components/component-instance.ts
@@ -1,0 +1,45 @@
+import type { FilePath } from '../file-entities/file.ts';
+import type { Position, PositionPath } from '../file-entities/position.ts';
+import type { ImportPath } from '../js-entities/import.ts';
+import type { Props } from '../jsx-entities/prop.ts';
+import type { ComponentId } from './component-id.ts';
+import type { ComponentName } from './component-name.ts';
+
+export type ComponentInstance = {
+  type: 'instance';
+  componentName: ComponentName;
+  componentId: ComponentId;
+  importedFrom?: ImportPath;
+  filePath: FilePath;
+  location: PositionPath;
+  isSelfClosing: boolean;
+  props: Props;
+  startPosition: Position;
+  endPosition: Position;
+};
+type CreateComponentInstanceArgs = Omit<ComponentInstance, 'type'>;
+
+export function createComponentInstance({
+  componentId,
+  componentName,
+  filePath,
+  importedFrom,
+  isSelfClosing,
+  location,
+  props,
+  startPosition,
+  endPosition,
+}: CreateComponentInstanceArgs): ComponentInstance {
+  return {
+    type: 'instance',
+    componentName,
+    componentId,
+    importedFrom,
+    filePath,
+    location,
+    isSelfClosing,
+    props,
+    startPosition,
+    endPosition,
+  };
+}
