+++ packages/jsx-scanner/src/components/component-name.ts
@@ -1,0 +1,1 @@
+export type ComponentName = string;
