+++ packages/jsx-scanner/src/main.ts
@@ -1,2 +1,5 @@
-export { type ComponentDefinition, type ComponentId, type ComponentInstance, type ComponentName } from './component.ts';
+export { type ComponentDefinition } from './components/component-definition.ts';
+export { type ComponentId } from './components/component-id.ts';
+export { type ComponentInstance } from './components/component-instance.ts';
+export { type ComponentName } from './components/component-name.ts';
 export { jsxScanner, type JsxScannerDiscovery } from './scanner.ts';
