+++ packages/jsx-scanner/src/parsers/assign-parser.ts
@@ -7,7 +7,9 @@ import {
   type SourceFile,
   type TypeChecker,
 } from 'typescript';
-import { type ComponentName, createComponentDefinition, getComponentId } from '../component.ts';
+import { createComponentDefinition } from '../components/component-definition.ts';
+import { getComponentId } from '../components/component-id.ts';
+import { type ComponentName } from '../components/component-name.ts';
 import { getRelativeFilePath } from '../file-entities/file.ts';
 import { getPosition, getPositionPath } from '../file-entities/position.ts';
 import type { GivenName } from '../js-entities/declaration.ts';
