--- packages/jsx-scanner/src/unique-id.ts
@@ -1,9 +1,0 @@
-import { createHash } from 'crypto';
-
-export type UniqueId = string;
-
-export function createUniqueId(string: string): UniqueId {
-  const hash = createHash('sha1').update(string);
-
-  return hash.digest('hex');
-}
